HSpec up server tests
